Migrate widgets to their own class
